[skip ci] Move Demo & Examples to be more obvious.
The live demo and API documentation links get lost at the top of the document. The first expectations in *most* GitHub repositories is an explanation of what the repository is, not links somewhere else. 

These are important links (especially for newcomers to the project) and not easy to find when scanning the Readme. 

Moving them to their own section improves their findabilty.